commit to pull new database

+++ GroupAssignment/Models/clsInvoice.cs
@@ -4,6 +4,6 @@ namespace GroupAssignment.Models {
     internal class Invoice {
         public int InvoiceNumber { get; set; }
         public DateTime InvoiceDate { get; set; }
-        public decimal TotalCost { get; set; }
+        public int TotalCost { get; set; }
     }
 }
